parser/requirements.txt
Pillow>=9.0.0
requests>=2.28.0
pytz>=2022.1
beautifulsoup4>=4.11.0
lxml>=4.9.0
pytesseract>=0.3.10
playwright>=1.40.0
playwright-stealth>=1.0.6
google-generativeai>=0.4.0
